Style creations; Background added; Buttons revised
OK, so I've made a couple more styles (one for button and the other for
the main layout [Background]). If you find the default BG ugly, won't be
a big deal to change. If you want to change any values that pertain to
BG, buttons, colors, and the like, modify styles.xml.